+++ src/FeatureInfo.js
@@ -1,6 +1,44 @@
 //import 'opening_hours';
+//declare module 'opening_hours';
 import opening_hours from 'opening_hours';
-//opening_hours.
+import titleCase from 'better-title-case';
+//const titleCase = require('better-title-case');
+
+let tagToIgnore=
+  '@id,name,NAME,NameS,addr:housename,addr:street,addr:city,addr:country,level,description,wheelchair,wheelchair:description,operator,' +
+  'opening_hours,access,centralkey,fee,fee:ChannelMergerNode,amenity,toilets:wheelchair,' +
+  'OBJECTID,Opening,UPRN,lon,lat,Address,Y,X,altitudeMode,OBJECTI,key,addr:housenumber,addr:postcode' +
+  'website,email,phone,source:addr' +
+  ',building,toilets:disposal,addr:city,addr:country,kitchen_hours,fhrs:id,layer,dataSupplier,dataLastUpdated'.split(',');
+  
+let dataImprover={
+  "capacity:disabled":"Number of Blue Badge spaces",
+  "capacity:parent":"Number of Parent and Child spaces",
+  "fee:amount:per_hour":"Carpark Charge (per hour)",
+  "fee:amount":"Carpark Charge",
+  "centralkey":"RADAR Key Required",
+  "access":"Access",
+  "kitchen_hours":"Kitchen Opening Times",
+  "opening_hours":"Opening Times",
+  "addr:housenumber":"",
+  "addr:street":"",
+  "addr:postcode":"",
+  "level":"Level",
+  "description":"",
+  "wheelchair:description":"",
+  "wheelchair":{label:"",
+                value:function(tagValue){
+                  switch (tagValue) {
+                    case "no":return "NO WHEELCHAIR ACCESS";
+                    case "yes":return "Wheelchair accessible";
+                    case "limited":return "Limited wheelchair access";
+                    default:return '';
+                  }
+                }},
+    'toilets:wheelchair':'Wheelchair Accessible Toilets',
+    'fee:charge':'Fee',
+    'operator':'Operated By',
+}
 
 function getCurrentlyOpen(featureOpeningHours){
   try {
@@ -33,6 +71,13 @@ function getCurrentlyOpen(featureOpeningHours){
 export default class FeatureInfo{
     constructor(properties,featureTags){
       this.tags=properties;
+      //few adhoc mappings to fix up the safePlaces data
+      if (this.tags.Opening) {this.tags.opening_hours=this.tags.Opening;}
+      if (this.tags.Name) {
+        this.tags.name=this.tags.Name;
+      }
+      if (this.tags.Address) {this.tags['addr:street']=this.tags.Address;}
+
       this.featureTags=featureTags.split(',');
     }
   
@@ -41,7 +86,7 @@ export default class FeatureInfo{
     }
   
     //return a labelled value if the value is valid, nothing otherwise
-    label(labelName,valueName, postValue,postLabel){
+    label(valueName, postValue, postLabel){
   
       let value='';
   
@@ -51,27 +96,48 @@ export default class FeatureInfo{
         
       }
       
-      let ret= 
-        (value)?
-          ( ((labelName)?`<strong>${labelName}:</strong>&nbsp;` + (postLabel||''):'') + value+ ((postValue)?postValue:'')):
-          '';
-      
+      if (!value) { return '';   } 
+      //look for any overrides or improvements and apply them if found
+      let dataImprovement =dataImprover[valueName];
+      let prettyLabel,prettyValue;
+      let improverType=typeof(dataImprovement);
+      switch (typeof(dataImprovement)){
+          case "string":
+              prettyLabel=dataImprovement;
+              prettyValue=value;
+              break;
+          case "object":
+              prettyLabel=dataImprovement.label;
+              prettyValue=dataImprovement.value(value);
+              break;
+          default:
+              prettyLabel = titleCase( valueName.replace(/[:|-|_]/gi,' '));
+              prettyValue=value;
+      }
+
+     //console.log(prettyLabel,prettyValue);
+      let ret = 
+           ((prettyLabel)?`<strong>${prettyLabel}:</strong>&nbsp;` + (postLabel ||''):'') + 
+           prettyValue + 
+           ((postValue)?postValue:'');
+          
       return ret;
     }
-  
+    
     get caption(){
-      return this.tags.name || this.tags["addr:housename"] || this.tags["@id"];
+      return this.tags.name || this.tags["addr:housename"] || this.tags.amenity || this.tags.building || this.tags["@id"];
     }
   
+    
     get location(){
-      return `${this.label('','addr:street')} ${this.label('Level','level')}<br/>
-              ${this.label('','description','<br/>')}
-              ${this.wheelchairAccessDesc}
-              ${this.label('','wheelchair:description')}
-      `;
+
+      return `${this.label('addr:housenumber',' ')} ${this.label('addr:street', ' ')} ${this.label('level',' ')}${this.label('addr:postcode',' ')}<br />
+              ${this.label('description','<br />')}
+              ${this.label('wheelchair', ' ')}
+              ${this.label('wheelchair:description')}`;
   
     }
-  
+  /*
     get wheelchairAccessDesc(){
       switch (this.tags.wheelchair) {
         case "no":return "NO WHEELCHAIR ACCESS";
@@ -79,51 +145,73 @@ export default class FeatureInfo{
         case "limited":return "Limited wheelchair access";
         default:return '';
       }
-    }
+    }*/
     get operator(){
-      return this.label('Operated By','operator', undefined,'<br/>') || "<strong>Information supplied by:</strong>Open Street Map";
+      return this.label('operator', '<br />');// || "<strong>Information supplied by:</strong>Open Street Map";
     }
+    
     get report(){
-      
-      let contact="";
-      
-      //if (supplier=osm)
-      let id=this.tags["@id"];
-      return `<a href='http://www.openstreetmap.org/edit?${id}' data-toggle='tooltip' title='Edit on OSM'>Update this information yourself</a>`
-      //return `<a href=''>Report Problem</a>`;
+      return this.label('website', '<br />') + 
+        this.label('email', '<br />') +
+        this.label('phone', '<br />');
+
     }
+    
+    get contact(){
+
+    }
+
+    //use the layer meta data to show the originator of the data and when it was last updated
+    get meta(){
+
+      let supplier="Unknown", lastUpdated="Unknown";
+      if (this.tags["@id"]) {
+        let id=this.tags["@id"].replace('/','=');
+        supplier= `<a href='http://www.openstreetmap.org/edit?${id}&comment=TCSM' target='_blank' data-toggle='tooltip' title='Edit on OSM'>Open Street Map</a>`
+      }
+      lastUpdated=this.tags.dataLastUpdated;
+      return `<strong>Data From:</strong> ${supplier} <strong>on:</strong> ${lastUpdated}`;
+    }
+
+    /*Use this section to list any conditions or barriers to accessing this service */
     get availability(){
 
       
-      let opening=this.label('Opening Times','opening_hours');
+      let opening=this.label('opening_hours');
       if(this.tags.opening_hours){
         let currentStatus = getCurrentlyOpen(this.tags.opening_hours);
         if (currentStatus){
           opening+=`<div class="pop-caption">${currentStatus}</div>`;
         } else {
-        opening+='<br/>';        
+        opening+='<br />';        
         }
       }
+      opening+=this.label('kitchen_hours');
            
       return opening + 
       //this.label('Opening Times','opening_hours',(currentStatus)?`<div class="pop-caption">${currentStatus}</div>`:'</br>') +
-      this.label('Access','access', '&nbsp;') +
-      this.label('RADAR key','centralkey', '&nbsp;') +
-      this.label('Fee','fee', this.tags["fee:charge"]) +
-      ((this.tags.amenity!=='toilets')?this.label('Wheelchair Accessible Toilets','toilets:wheelchair'):'');
+      this.label('access', '&nbsp;') +
+      this.label('centralkey', '&nbsp;') +
+      this.label('fee:charge') +
+      ((this.tags.amenity!=='toilets')?this.label('toilets:wheelchair'):'');
       
     }
+
+    
 
   
 
     get facilities(){
-      return (
-      this.label('Changing Table','changing_table',"</br>") +
-      this.label('Unisex','unisex',"</br>") +
-      this.label('Female','female',"</br>") +
-      this.label('Male','male',"</br>") +
-      this.label('Drinking Water','drinking_water'));
-      //return ret;
+      //return any fields not shown elsewhere
+      let facList= Object.entries(this.tags).reduce((acc,[key,value]) => {
+        if (!tagToIgnore.includes(key)) {
+          //let keyLabel =key.replace(/[:-]/gi,' ');
+          acc.push( this.label(key));
+        }
+        return acc;
+        },[]  ); 
+      return facList.join('<br />');
+
     }
   
     get rawData(){