+++ src/better-title-case.d.ts
@@ -1,0 +1,1 @@
+declare module 'better-title-case';